requirements.txt
# Enigma OS — Phase 1 Foundation
# Notion + Supabase + Python orchestration

notion-client>=2.2.1
supabase>=2.3.0
python-dotenv>=1.0.0
cryptography>=42.0.0
openai>=1.0.0
fastapi>=0.108.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
PyJWT[crypto]>=2.8.0
